Allow direct use of archive URLs in harness attribute
