+++ config/services.yml
@@ -47,6 +47,14 @@ services:
   my127\Console\:
     resource: '../vendor/my127/my127/src/Console/src/*'
 
+  my127\Workspace\Types\Harness\Repository\AggregateRepository: ~
+  my127\Workspace\Types\Harness\Repository\ArchiveRepository: ~
+  my127\Workspace\Types\Harness\Repository\PackageRepository: ~
+
+  my127\Workspace\Types\Harness\Repository\Repository $packages: '@my127\Workspace\Types\Harness\Repository\AggregateRepository'
+  my127\Workspace\Types\Harness\Repository\Repository $archiveRepository: '@my127\Workspace\Types\Harness\Repository\ArchiveRepository'
+  my127\Workspace\Types\Harness\Repository\Repository $packageRepository: '@my127\Workspace\Types\Harness\Repository\PackageRepository'
+
   # -- require
 
   Symfony\Component\EventDispatcher\EventDispatcher:
